+++ shared/tasks/scripts.mjs
@@ -103,7 +103,8 @@ export function compile(
       file: !output.preserveModules ? join(destPath, outputPath) : undefined,
 
       // Enable source maps
-      sourcemap: true
+      sourcemap: true,
+      sourcemapExcludeSources: !!output.preserveModules
     })
   })
 }