+++ shared/tasks/styles.mjs
@@ -38,7 +38,8 @@ export function compile(inputPath, { srcPath, destPath, output = {} }) {
     map: /** @type {ProcessOptions['map']} */ ({
       annotation: true,
       inline: false,
-      prev: false
+      prev: false,
+      sourcesContent: false
     }),
 
     // Sass syntax support
@@ -73,6 +74,7 @@ export function compile(inputPath, { srcPath, destPath, output = {} }) {
 
         // Pass source maps to PostCSS
         options.map.prev = map
+        options.map.sourcesContent = true
       }
     }